very broken function moves
