+++ isofit/radiative_transfer/luts.py
@@ -962,3 +962,172 @@ def cleanup(file):
                     f"The LUT status was determined to be incomplete, auto-removing: {file}"
                 )
                 os.remove(file)
+
+
+# All functions need to resolve all calls
+
+
+def runSimulations(self) -> None:
+    """
+    Run all simulations for the LUT grid.
+
+    """
+    Logger.info(f"Running any pre-sim functions")
+    pre = self.preSim()
+
+    if pre:
+        Logger.info("Saving pre-sim data to index zero of all dimensions except wl")
+        Logger.debug(f"pre-sim data contains keys: {pre.keys()}")
+
+        point = {key: 0 for key in self.lut_names}
+        self.lut.writePoint(point, data=pre)
+
+    # Make the LUT calls (in parallel if specified)
+    if not self._disable_makeSim:
+        Logger.info("Executing parallel simulations")
+
+        # Place into shared memory space to avoid spilling
+        lut_names = ray.put(self.lut_names)
+        makeSim = ray.put(self.makeSim)
+        readSim = ray.put(self.readSim)
+        lut_path = ray.put(self.lut_path)
+        buffer_time = ray.put(self.max_buffer_time)
+        rte_configure_and_exit = ray.put(self.engine_config.rte_configure_and_exit)
+
+        jobs = [
+            streamSimulation.remote(
+                point,
+                lut_names,
+                makeSim,
+                readSim,
+                lut_path,
+                max_buffer_time=buffer_time,
+                rte_configure_and_exit=self.engine_config.rte_configure_and_exit,
+            )
+            for point in self.points
+        ]
+
+        if self.engine_config.rte_configure_and_exit:
+            # Block until all jobs finish
+            ray.get(jobs)
+
+            Logger.warning("Exiting early due to rte_configure_and_exit")
+            sys.exit(0)
+        else:
+            # Report a percentage complete every 10% and flush to disk at those intervals
+            report = common.Track(
+                jobs,
+                step=10,
+                reverse=True,
+                print=Logger.info,
+                message="simulations complete",
+            )
+
+            # Update the lut as point simulations stream in
+            while jobs:
+                [done], jobs = ray.wait(jobs, num_returns=1)
+
+                # Retrieve the return of the finished job
+                ret = ray.get(done)
+
+                # If a simulation fails then it will return None
+                if ret:
+                    self.lut.queuePoint(*ret)
+
+                if report(len(jobs)):
+                    Logger.info("Flushing netCDF to disk")
+                    self.lut.flush()
+
+            # Shouldn't be hit but just in case
+            if self.lut.hold:
+                Logger.warning("Not all points were flushed, doing so now")
+                self.lut.flush()
+
+        del lut_names, makeSim, readSim, lut_path, buffer_time
+    else:
+        Logger.debug("makeSim is disabled for this engine")
+
+    Logger.info(f"Running any post-sim functions")
+    post = self.postSim()
+
+    if post:
+        Logger.info("Saving post-sim data to index zero of all dimensions except wl")
+        Logger.debug(f"post-sim data contains keys: {post.keys()}")
+
+        point = {key: 0 for key in self.lut_names}
+        self.lut.writePoint(point, data=post)
+
+    self.lut.finalize()
+
+    # Reload the LUT now that it's populated
+    Logger.debug("Reloading LUT")
+    self.lut = luts.load(self.lut_path)
+
+
+@ray.remote(num_cpus=1)
+def streamSimulation(
+    point: np.array,
+    lut_names: list,
+    simmer: Callable,
+    reader: Callable,
+    output: str,
+    max_buffer_time: float = 0.5,
+    rte_configure_and_exit: bool = False,
+):
+    """Run a simulation for a single point and stream the results to a saved lut file.
+
+    Args:
+        point (np.array): conditions to alter in simulation
+        lut_names (list): Dimension names aka lut_names
+        simmer (function): function to run the simulation
+        reader (function): function to read the results of the simulation
+        output (str): LUT store to save results to
+        max_buffer_time (float, optional): _description_. Defaults to 0.5.
+        rte_configure_and_exit (bool, optional): exit early if not executing simulations
+    """
+    Logger.debug(f"Simulating(point={point})")
+
+    # Slight delay to prevent all subprocesses from starting simultaneously
+    time.sleep(np.random.rand() * max_buffer_time)
+
+    # Execute the simulation
+    simmer(point)
+
+    # No data will be produced, just configuration files
+    if rte_configure_and_exit:
+        return
+
+    # Read the simulation results
+    data = reader(point)
+
+    # Save the results to our LUT format
+    if data:
+        Logger.debug(f"Updating data point {point} for keys: {data.keys()}")
+
+        return point, data
+    else:
+        Logger.warning(f"No data was returned for point {point}")
+
+
+def build_interpolators(self):
+    """
+    Builds the interpolators using the LUT store
+
+    TODO: optional load from/write to disk
+    """
+    self.luts = {}
+
+    ds = self.lut.unstack("point")
+
+    # Make sure its in expected order, wl at the end
+    ds = ds.transpose(*self.lut_names, "wl")
+
+    grid = [ds[key].data for key in self.lut_names]
+
+    # Create the unique
+    for key in luts.Keys.alldim:
+        self.luts[key] = common.VectorInterpolator(
+            grid_input=grid,
+            data_input=ds[key].load().data,
+            version=self.interpolator_style,
+        )
